smart-gateway/requirements.txt
fastapi
uvicorn[standard]
strawberry-graphql
grpcio
grpcio-tools
python-dotenv
langchain
langchain-openai
